validation: flush undo files when the highest height block undo data has been written
Data files are pre-allocated, and upon flush/finalization, they are trimmed down to their resulting size.
Block (blk) files are written to as blocks come in, which is often out of order, whereas undo (rev) files are written sequentially, as each block is added to the top of the chain.
When a block file hits the size limit, the system flushes and trims the file down to its final size, and moves on to the next block file.
At this point, there will be some blocks in the old block file that have yet to be connected. As a result, the old rev file will still be written to for awhile.
Specifically, the old rev file will be written to until we have written undo data for the highest height block in the old block file, which is given by the CBlockFileInfo.nHeightLast value.

The old code would erroneously flush the undo file alongside the block file, but the undo file would immediately be re-opened, and pre-allocation would occur again (up to 1 MB). In the worst case scenario, a rev file will end up with 1 MB of allocated but unused space. At this point in time, there are about two thousand rev files, which means roughly 2 GB of wasted space.

This commit adds a secondary flush/finalize call, triggered after the undo data for the highest height block in the corresponding block file has been written. This will trim away the surplus (up to 1 MB) space above.